+++ src/macros.rs
@@ -146,12 +146,7 @@ macro_rules! assert_toml_snapshot {
 /// just use an empty string (`@""`).
 ///
 /// The snapshot name is optional but can be provided as first argument.
-#[cfg(feature = "serde")]
-#[cfg_attr(
-    not(feature = "yaml"),
-    deprecated(note = "assert_yaml_snapshot! will require the \"yaml\" feature. \
-        Add the \"yaml\" feature to your Cargo.toml to silence this warning.")
-)]
+#[cfg(feature = "yaml")]
 #[cfg_attr(docsrs, doc(cfg(feature = "yaml")))]
 #[macro_export]
 macro_rules! assert_yaml_snapshot {
@@ -239,12 +234,7 @@ macro_rules! assert_ron_snapshot {
 /// about redactions refer to the [redactions feature in the guide](https://insta.rs/docs/redactions/).
 ///
 /// The snapshot name is optional but can be provided as first argument.
-#[cfg(feature = "serde")]
-#[cfg_attr(
-    not(feature = "json"),
-    deprecated(note = "assert_json_snapshot! will require the \"json\" feature. \
-        Add the \"json\" feature to your Cargo.toml to silence this warning.")
-)]
+#[cfg(feature = "json")]
 #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
 #[macro_export]
 macro_rules! assert_json_snapshot {
@@ -417,14 +407,6 @@ macro_rules! assert_snapshot {
         $crate::assert_snapshot!($name, $value, stringify!($value))
     };
     ($name:expr, $value:expr, $debug_expr:expr) => {{
-        // This use does not do anything.  It exists purely to trigger a deprecation warning
-        // for the features if enabled.
-        #[allow(unused)]
-        {
-            use $crate::deprecated_backtrace_support::*;
-            use $crate::deprecated_serialization_support::*;
-        }
-
         $crate::_macro_support::assert_snapshot(
             // Creates a ReferenceValue::Named variant
             $name.into(),