+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to insta and cargo-insta are documented here.
 
 ## 0.18.0
 
+*(unreleased)*
+
 - `Settings::bind` now can return a result.
 - Expose the drop guard type of `bind_to_scope`.
 - The `serde` dependency is now optional. While still enabled by default