+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to insta and cargo-insta are documented here.
 
+## 1.19.0
+
+- Removed `backtrace` feature.
+- Removed `serialization` feature.
+- `assert_json_snapshot!` and `assert_yaml_snapshot!` now require
+  the `json` and `yaml` feature respectively.
+
 ## 1.18.0
 
 - `Settings::bind` now can return a result.